governance/anonymizer/requirements.txt
apache-flink==1.19.0
pyflink==1.19.0
kafka-python==2.0.2
confluent-kafka==2.4.0
hvac==2.1.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4
pyyaml==6.0.1
cryptography==42.0.8
ff3==1.0.1
requests==2.32.3
